requirements.txt
# Browser automation
playwright>=1.40.0

# Image handling
pillow>=10.0.0

# Computer vision (cv2)
opencv-python>=4.8.0

# OCR wrapper
pytesseract>=0.3.10

# Numerical dependency (required by OpenCV)
numpy>=1.24.0

# PDF extraction (Phase 3) - install at least one
pdfplumber>=0.10.0
PyPDF2>=3.0.0

mysql-connector-python==8.3.0
